Module-4/backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==1.0.0
gunicorn==21.2.0
torch==2.7.0
torchvision==0.18.0
numpy==1.24.3
Pillow==10.2.0
opencv-python==4.9.0.80
scikit-learn==1.3.2
pandas==2.1.4
matplotlib==3.8.2
scipy==1.11.4
tqdm==4.66.1 